requirements.txt
amqp==2.6.1
asgiref==3.5.2
autopep8==1.7.0
Babel==2.11.0
billiard==3.6.4.0
braintree==3.59.0
cairocffi==1.4.0
CairoSVG==2.5.2
celery==4.4.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cssbeautifier==1.14.6
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2
django-braces==1.14.0
django-embed-video==1.3.2
django-extensions==2.2.5
django-faker==0.2
django-localflavor==3.0.1
django-memcache-status==2.2
django-parler==2.2.1
django-rosetta==0.9.3
django-taggit==1.2.0
djangorestframework==3.11.0
djlint==1.19.1
dnspython==2.2.1
easy-thumbnails==2.7
EditorConfig==0.12.3
eventlet==0.33.1
fake-factory==9999.9.9
flower==0.9.3
greenlet==2.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
jsbeautifier==1.14.6
kombu==4.6.11
Markdown==3.2.1
oauthlib==3.2.2
pathspec==0.10.1
Pillow==7.0.0
polib==1.1.1
psycopg2-binary==2.8.4
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyphen==0.13.0
python-memcached==1.59
python-stdnum==1.17
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0
redis==3.4.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==3.1.0
social-auth-core==4.3.0
sqlparse==0.4.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==5.1.1
tqdm==4.64.1
urllib3==1.26.12
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
Werkzeug==0.16.0
zipp==3.9.0
